Port to amd64 (note: needs ld-elf.so.1 bugfix to actually work)
I've taken a chance and split the sparc64 and alpha mega-diffs up because
sparc64 patched the result of alpha's patches, and amd64 needed to do the
same again - but it was sorted lexically before alpha/sparc64.  I've
grouped the xptcall patches together since they are standalone and split
up and merged the other infrastructure jumbo-patches.

I also attempted to sync up the nspr vs firebird diffs for sparc64 and
alpha because they disagreed.  HOWEVER... While this port 'depends'
on the nspr port, it doesn't actually *use* it.  It should probably
be removed from the LIB_DEPENDS list.

Note that the amd64 xptcall code came from SuSE via the mozilla bugzilla
database.