- Update to 1.0.1
Changes: This version is stable after a month of testing. Code clean-up
was done. No major bugs were found, but a minor bug was corrected. The
access mode was changed from +o to +f. Multiple or shared database files
for each channel and major updates on the documentation files. This code
is final unless some major bug is reported or huge improvements are made.