Change the default version of ruby to 1.8 for i386 as well, finally.
Always put a version suffix to the ruby name (no matter if ruby is the
default version) to avoid mess in future.

[Notes for i386 users]

If you are a ruby developer and still want to stick with ruby 1.6 as
default, please add RUBY_DEFAULT_VER=1.6 to /etc/make.conf.

If you are a ruby developer and want to keep ruby 1.6 as default,
please add RUBY_DEFAULT_VER=1.6 to /etc/make.conf.  Otherwise, please
run the following series of commands to migrate to ruby 1.8:

  1) Reinstall portupgrade manually (and ruby 1.8 will be installed)
	pkg_delete portupgrade-\*
        (cd /usr/ports/sysutils/portupgrade; make install clean)

  2) Reinstall everything that depends on ruby 1.6 (to use ruby 1.8)
	portupgrade -fr lang/ruby16

  3) Reinstall ruby 1.8 (because the previous step kills symlinks)
	portupgrade -f lang/ruby18

  4) Deinstall ruby 1.6 stuff (if you are paranoia)
        pkg_deinstall -ri lang/ruby16