Upgrade to version 3.0.11b8.
The only difference to the previous version (except for the version ID
sent in request packets), is the method used to declare errno.

For unknown reasons, this version is not available from all mirror sites
and binary packages seem to only be provided for 3.0.11b7. The problem
with that version was that it was used incorrectly in some Linksys WLAN
router which tended to spam DynDNS.org with bogus requests. DynDNS.org
ignores certain requests, when received with a version ID of 3.0.11b7,
though there's nothing wrong with that code when used correctly.

I had considered to just patch the version string in "configure", but
after some more thought I've now choosen to actually use the new dist
file and to remove the URL of the master site that doesn't offer the
newer version.

PR:		60940
Submitted by:	Lynk <lynk@lynksys.net>