Update libgcrypt to the new stable version 1.2.0 Retire libgcrypt-devel
Switch back to security/libgcrypt and bump PORTREVISION for
- gnutls
- gnutls-devel
- gsasl
- opencdk
- vpnc

libggz:
- Remove dependency on libgcrypt, because the port does not yet
work with the new libgcrypt

libksba:
- Update to 0.9.5
- Pass maintainership to lofi, who already maintains the other
aegypten ports.

Approved by:	lofi

wmbiff:
Switch to security/libgcrypt but no PORTREVISION bump, because the
dependency is optional and the port does not work with the new
libgcrypt and needs to be updated.

No action, because ports don't work with the new libgcrypt and need
to be updated, maintainers informed:
security/newpg
security/pinentry
security/dirmngr
emulators/fuse
emulators/libspectrum