Upgrade to 0.60.1.
This release adds a new Nroff and Texinfo filter and fixes a number
of bugs and portability problems.

Warning: aspell-0.60.1 is now incompatible with the base ncurses.
To keep the ncurses interfaces, please define WITH_NCURSES: aspell
will then depend on the port devel/ncurses.
Since devel/ncurses is not packageable, this is not the default.