Nautilus-cd-burner now tries to use UTF-8 encoding when creating ISO images. This character set is only available in the cdrtools-cjk slave port, so depend on that now.
This is not an ideal solution since users of cdrtools will see an infinite
loop when trying to use nautilus-cd-burner, and portupgrade will not handle
this transition smoothly.  I'll try to see if marius would be willing to
install the cdrtools-cjk binaries with a different name so that the two
ports will not conflict.